+++ lib/View/Screens/intro_screens/screen1.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+class HomePage extends StatelessWidget {
+  const HomePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        flexibleSpace: Container(
+          decoration: const BoxDecoration(
+              gradient: LinearGradient(
+                  colors: [Colors.indigoAccent, Colors.purpleAccent])),
+        ),
+      ),
+    );
+  }
+}
